Renamed "styles" folder to "resources"; renamed "common.dart" to "common_widgets.dart" and moved to "resources" folder.
