+++ lib/main.dart
@@ -9,7 +9,7 @@ import 'package:athlete_surveyor/models/inbox_model.dart';
 import 'package:athlete_surveyor/models/previous_forms_model.dart';
 import 'package:athlete_surveyor/pages/inbox_page.dart';
 import 'package:athlete_surveyor/pages/previous_forms_page.dart';
-import 'package:athlete_surveyor/styles/colors.dart';
+import 'package:athlete_surveyor/resources/colors.dart';
 import 'package:flutter/material.dart';
 
 const Text _appbarTitleText = Text(
